Separate package for RedisBloom commands
