+++ src/main/java/redis/clients/jedis/bloom/commands/BloomFilterPipelineCommands.java
@@ -1,8 +1,10 @@
-package redis.clients.jedis.bloom;
+package redis.clients.jedis.bloom.commands;
 
 import java.util.List;
 import java.util.Map;
 import redis.clients.jedis.Response;
+import redis.clients.jedis.bloom.BFInsertParams;
+import redis.clients.jedis.bloom.BFReserveParams;
 
 public interface BloomFilterPipelineCommands {
