+++ src/main/java/redis/clients/jedis/bloom/commands/RedisBloomCommands.java
@@ -1,4 +1,4 @@
-package redis.clients.jedis.bloom;
+package redis.clients.jedis.bloom.commands;
 
 public interface RedisBloomCommands extends BloomFilterCommands, CuckooFilterCommands,
     CountMinSketchCommands, TopKFilterCommands, TDigestSketchCommands {
