+++ src/main/java/redis/clients/jedis/bloom/commands/RedisBloomPipelineCommands.java
@@ -1,4 +1,4 @@
-package redis.clients.jedis.bloom;
+package redis.clients.jedis.bloom.commands;
 
 public interface RedisBloomPipelineCommands extends BloomFilterPipelineCommands,
     CuckooFilterPipelineCommands, CountMinSketchPipelineCommands, TopKFilterPipelineCommands,
