+++ src/main/java/redis/clients/jedis/bloom/commands/TDigestSketchCommands.java
@@ -1,7 +1,8 @@
-package redis.clients.jedis.bloom;
+package redis.clients.jedis.bloom.commands;
 
 import java.util.List;
 import java.util.Map;
+import redis.clients.jedis.bloom.TDigestMergeParams;
 
 public interface TDigestSketchCommands {
