+++ src/main/java/redis/clients/jedis/bloom/commands/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package contains the interfaces that contain methods representing RedisBloom commands.
+ */
+package redis.clients.jedis.bloom.commands;
