+++ src/main/java/redis/clients/jedis/bloom/package-info.java
@@ -1,4 +1,4 @@
 /**
- * This package contains the classes and interfaces related to RedisBloom module.
+ * This package contains the classes related to RedisBloom module.
  */
 package redis.clients.jedis.bloom;
