+++ src/main/java/redis/clients/jedis/commands/RedisModuleCommands.java
@@ -1,6 +1,6 @@
 package redis.clients.jedis.commands;
 
-import redis.clients.jedis.bloom.RedisBloomCommands;
+import redis.clients.jedis.bloom.commands.RedisBloomCommands;
 import redis.clients.jedis.graph.RedisGraphCommands;
 import redis.clients.jedis.json.RedisJsonCommands;
 import redis.clients.jedis.search.RediSearchCommands;
