+++ src/main/java/redis/clients/jedis/commands/RedisModulePipelineCommands.java
@@ -1,6 +1,6 @@
 package redis.clients.jedis.commands;
 
-import redis.clients.jedis.bloom.RedisBloomPipelineCommands;
+import redis.clients.jedis.bloom.commands.RedisBloomPipelineCommands;
 import redis.clients.jedis.graph.RedisGraphPipelineCommands;
 import redis.clients.jedis.json.RedisJsonPipelineCommands;
 import redis.clients.jedis.search.RediSearchPipelineCommands;
